hub/requirements.txt
# Home Server Hub Requirements
flask>=3.0.0
python-dotenv>=1.0.0
psutil>=5.9.0
pyyaml>=6.0.0
